Several upgrade for WZ VFF:
Core changes: I added a sub to solve the bulk_material for a couple of atoms. Material is parsed depending on who's an anion.
This checked is hardcoded (TODO: implement a generic version based on database infos).
The tricky hardcoded part is now completely in AtomisticStructure.
An hardcodedcheck on the "c-direction" is in VffModel, this must be generalized (but the check has been wrapped in a sub call)

User can set his own parameters in the input file.